TM-19 change repository type from Long to Integer

+++ src/main/java/com/tma/teamhr/repository/IArticleRepository.java
@@ -3,5 +3,5 @@ package com.tma.teamhr.repository;
 import com.tma.teamhr.model.Article;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IArticleRepository extends CrudRepository<Article, Long> {
+public interface IArticleRepository extends CrudRepository<Article, Integer> {
 }
