+++ src/main/java/com/tma/teamhr/repository/IColumnRepository.java
@@ -3,5 +3,5 @@ package com.tma.teamhr.repository;
 import com.tma.teamhr.model.Columnn;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IColumnRepository extends CrudRepository<Columnn, Long> {
+public interface IColumnRepository extends CrudRepository<Columnn, Integer> {
 }
